requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
abseil-cpp=20211102.0=h27087fc_1
absl-py=1.3.0=pypi_0
aiohttp=3.8.3=pypi_0
aiosignal=1.3.1=pypi_0
appdirs=1.4.4=pyhd3eb1b0_0
arrow-cpp=8.0.0=py39h60b952e_0
asttokens=2.1.0=pypi_0
async-timeout=4.0.2=pypi_0
attrs=22.1.0=pypi_0
aws-c-common=0.4.57=he6710b0_1
aws-c-event-stream=0.1.6=h2531618_5
aws-checksums=0.1.9=he6710b0_0
aws-sdk-cpp=1.8.185=hce553d0_0
backcall=0.2.0=pypi_0
blas=1.0=mkl
blosc=1.21.0=h4ff587b_1
boost-cpp=1.78.0=he72f1d9_0
brotli=1.0.9=h166bdaf_7
brotli-bin=1.0.9=h166bdaf_7
brotlipy=0.7.0=py39h27cfd23_1003
brunsli=0.1=h2531618_0
bzip2=1.0.8=h7b6447c_0
c-ares=1.18.1=h7f98852_0
ca-certificates=2023.01.10=h06a4308_0
cachetools=5.2.0=pypi_0
certifi=2022.12.7=py39h06a4308_0
cffi=1.15.1=py39h5eee18b_2
cfitsio=3.470=h5893167_7
charls=2.2.0=h2531618_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.3=pypi_0
cloudpickle=2.0.0=pyhd3eb1b0_0
colorama=0.4.6=pypi_0
contourpy=1.0.6=pypi_0
cryptography=38.0.1=py39h9ce1e76_0
cuda=11.6.2=0
cuda-cccl=11.6.55=hf6102b2_0
cuda-command-line-tools=11.6.2=0
cuda-compiler=11.6.2=0
cuda-cudart=11.6.55=he381448_0
cuda-cudart-dev=11.6.55=h42ad0f4_0
cuda-cuobjdump=11.6.124=h2eeebcb_0
cuda-cupti=11.6.124=h86345e5_0
cuda-cuxxfilt=11.6.124=hecbf4f6_0
cuda-driver-dev=11.6.55=0
cuda-gdb=11.8.86=0
cuda-libraries=11.6.2=0
cuda-libraries-dev=11.6.2=0
cuda-memcheck=11.8.86=0
cuda-nsight=11.8.86=0
cuda-nsight-compute=11.8.0=0
cuda-nvcc=11.6.124=hbba6d2d_0
cuda-nvdisasm=11.8.86=0
cuda-nvml-dev=11.6.55=haa9ef22_0
cuda-nvprof=11.8.87=0
cuda-nvprune=11.6.124=he22ec0a_0
cuda-nvrtc=11.6.124=h020bade_0
cuda-nvrtc-dev=11.6.124=h249d397_0
cuda-nvtx=11.6.124=h0630a44_0
cuda-nvvp=11.8.87=0
cuda-runtime=11.6.2=0
cuda-samples=11.6.101=h8efea70_0
cuda-sanitizer-api=11.8.86=0
cuda-toolkit=11.6.2=0
cuda-tools=11.6.2=0
cuda-visual-tools=11.6.2=0
cycler=0.11.0=pypi_0
cytoolz=0.12.0=py39h5eee18b_0
dask-core=2022.7.0=py39h06a4308_0
decorator=5.1.1=pypi_0
docker-pycreds=0.4.0=pypi_0
docopt=0.6.2=pypi_0
einops=0.3.0=pypi_0
executing=1.2.0=pypi_0
ffmpeg=4.3=hf484d3e_0
fftw=3.3.9=h27cfd23_1
filelock=3.8.0=pypi_0
fonttools=4.38.0=pypi_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.3.3=pypi_0
fsspec=2022.11.0=py39h06a4308_0
ftfy=6.0.3=pypi_0
future=0.18.2=pypi_0
gds-tools=1.4.0.31=0
gflags=2.2.2=he1b5a44_1004
giflib=5.2.1=h7b6447c_0
gitdb=4.0.10=pypi_0
gitpython=3.1.29=pypi_0
glog=0.6.0=h6f12383_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
google-auth=2.14.1=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpc-cpp=1.46.1=h33aed49_0
grpcio=1.50.0=pypi_0
huggingface-hub=0.0.8=pypi_0
icu=70.1=h27087fc_0
idna=3.4=py39h06a4308_0
imagecodecs=2021.8.26=py39hf0132c2_1
imageio=2.19.3=py39h06a4308_0
imbalanced-learn=0.10.1=pypi_0
importlib-metadata=5.1.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
ipdb=0.13.4=pypi_0
ipython=8.6.0=pypi_0
jedi=0.18.2=pypi_0
joblib=1.2.0=pypi_0
jpeg=9e=h7f8727e_0
jsonpickle=1.5.2=pypi_0
jxrlib=1.1=h7b6447c_2
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.4=pypi_0
krb5=1.19.3=h3790be6_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libaec=1.0.4=he6710b0_1
libbrotlicommon=1.0.9=h166bdaf_7
libbrotlidec=1.0.9=h166bdaf_7
libbrotlienc=1.0.9=h166bdaf_7
libcublas=11.11.3.6=0
libcublas-dev=11.11.3.6=0
libcufft=10.9.0.58=0
libcufft-dev=10.9.0.58=0
libcufile=1.4.0.31=0
libcufile-dev=1.4.0.31=0
libcurand=10.3.0.86=0
libcurand-dev=10.3.0.86=0
libcurl=7.85.0=h91b91d3_0
libcusolver=11.4.1.48=0
libcusolver-dev=11.4.1.48=0
libcusparse=11.7.5.86=0
libcusparse-dev=11.7.5.86=0
libdeflate=1.8=h7f8727e_5
libedit=3.1.20191231=he28a2e2_2
libev=4.33=h516909a_1
libevent=2.1.10=h9b69904_4
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libnghttp2=1.46.0=hce63b2e_0
libnpp=11.8.0.86=0
libnpp-dev=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libnvjpeg-dev=11.9.0.86=0
libpng=1.6.37=hbc83047_0
libprotobuf=3.20.1=h4ff587b_0
libssh2=1.10.0=ha56f1ee_2
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.16.0=h27cfd23_0
libthrift=0.15.0=he6d91bd_0
libtiff=4.4.0=hecacb30_2
libunistring=0.9.10=h27cfd23_0
libwebp=1.2.4=h11a3e52_0
libwebp-base=1.2.4=h5eee18b_0
libzopfli=1.0.3=he6710b0_0
locket=1.0.0=py39h06a4308_0
lz4-c=1.9.3=h295c915_1
markdown=3.4.1=pypi_0
markupsafe=2.1.1=pypi_0
matplotlib=3.6.2=pypi_0
matplotlib-inline=0.1.6=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py39h7f8727e_0
mkl_fft=1.3.1=py39hd3c417c_0
mkl_random=1.2.2=py39h51133e4_0
multidict=6.0.2=pypi_0
munch=2.5.0=pypi_0
ncurses=6.3=h5eee18b_3
nettle=3.7.3=hbbd107a_1
networkx=2.8.4=py39h06a4308_0
nsight-compute=2022.3.0.22=0
numpy=1.23.4=py39h14f4228_0
numpy-base=1.23.4=py39h31eccc5_0
oauthlib=3.2.2=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openssl=1.1.1s=h7f8727e_0
orc=1.7.4=h07ed6aa_0
packaging=21.3=pypi_0
pandas=1.1.5=pypi_0
parso=0.8.3=pypi_0
partd=1.2.0=pyhd3eb1b0_1
pathtools=0.1.2=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=8.1.0=pypi_0
pip=22.2.2=py39h06a4308_0
pooch=1.4.0=pyhd3eb1b0_0
promise=2.3=pypi_0
prompt-toolkit=3.0.33=pypi_0
protobuf=3.20.3=pypi_0
psutil=5.9.4=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
py-cpuinfo=9.0.0=pypi_0
pyarrow=8.0.0=py39h992f0b0_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydeprecate=0.3.0=pypi_0
pygments=2.13.0=pypi_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=pypi_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.15=h7a1cb2a_2
python-dateutil=2.8.2=pypi_0
pytorch=1.13.0=py3.9_cuda11.6_cudnn8.3.2_0
pytorch-cuda=11.6=h867d48c_0
pytorch-lightning=1.3.2=pypi_0
pytorch-mutex=1.0=cuda
pytz=2022.6=pypi_0
pywavelets=1.4.1=py39h5eee18b_0
pyyaml=5.4.1=pypi_0
re2=2022.04.01=h27087fc_0
readline=8.2=h5eee18b_0
regex=2022.10.31=pypi_0
requests=2.28.1=py39h06a4308_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
sacred=0.8.2=pypi_0
sacremoses=0.0.53=pypi_0
scikit-image=0.19.3=py39h6a678d5_1
scikit-learn=1.0.2=pypi_0
scipy=1.9.3=pypi_0
sentry-sdk=1.11.1=pypi_0
setproctitle=1.3.2=pypi_0
setuptools=65.5.0=py39h06a4308_0
shortuuid=1.0.11=pypi_0
six=1.16.0=pyhd3eb1b0_1
smmap=5.0.0=pypi_0
snappy=1.1.9=hbd366e4_1
sqlite=3.40.0=h5082296_0
stack-data=0.6.1=pypi_0
tensorboard=2.11.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
threadpoolctl=3.1.0=pypi_0
tifffile=2021.7.2=pyhd3eb1b0_2
timm=0.4.12=pyhd8ed1ab_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.10.3=pypi_0
toolz=0.12.0=py39h06a4308_0
torchaudio=0.13.0=py39_cu116
torchmetrics=0.7.2=pypi_0
torchvision=0.14.0=py39_cu116
tqdm=4.56.0=pypi_0
traitlets=5.5.0=pypi_0
transformers=4.6.0=pypi_0
typing_extensions=4.3.0=py39h06a4308_0
tzdata=2022f=h04d1e81_0
urllib3=1.26.12=py39h06a4308_0
utf8proc=2.6.1=h27cfd23_0
wandb=0.13.5=pypi_0
wcwidth=0.2.5=pypi_0
werkzeug=2.2.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
wrapt=1.14.1=pypi_0
xz=5.2.6=h5eee18b_0
yaml=0.2.5=h7b6447c_0
yarl=1.8.1=pypi_0
zfp=0.5.5=h295c915_6
zipp=3.11.0=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.2=ha4553b6_0
